requirements.txt
blinker==1.7.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
filelock==3.13.4
Flask==3.0.3
fsspec==2024.3.1
gevent==24.2.1
greenlet==3.0.3
huggingface-hub==0.22.2
idna==3.7
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
opencv-python==4.9.0.80
packaging==24.0
pillow==10.3.0
pycparser==2.22
PyYAML==6.0.1
requests==2.31.0
safetensors==0.4.2
sympy==1.12
timm==0.9.16
torch==2.2.2
torchvision==0.17.2
tqdm==4.66.2
typing_extensions==4.11.0
urllib3==2.2.1
Werkzeug==3.0.2
zope.event==5.0
zope.interface==6.3

